characteristic/u2ac3/src/main/java/org/im97mori/ble/characteristic/u2ac3/ObjectIdAndroid.java
package org.im97mori.ble.characteristic.u2ac3;

import android.bluetooth.BluetoothGattCharacteristic;
import android.os.Parcel;
import android.os.Parcelable;

import androidx.annotation.NonNull;

import org.im97mori.ble.ByteArrayCreator;

import java.util.Objects;

/**
 * object ID (Characteristics UUID: 0x2AC3)
 */
@SuppressWarnings({"WeakerAccess"})
public class ObjectIdAndroid extends ObjectId implements Parcelable {

    /**
     * @see ByteArrayCreator
     */
    public static final ByteArrayCreator<ObjectIdAndroid> CREATOR = new ByteArrayCreator<ObjectIdAndroid>() {

        /**
         * {@inheritDoc}
         */
        @Override
        @NonNull
        public ObjectIdAndroid createFromParcel(@NonNull Parcel in) {
            return new ObjectIdAndroid(in);
        }

        /**
         * {@inheritDoc}
         */
        @Override
        @NonNull
        public ObjectIdAndroid[] newArray(int size) {
            return new ObjectIdAndroid[size];
        }

        /**
         * {@inheritDoc}
         */
        @NonNull
        public ObjectIdAndroid createFromByteArray(@NonNull byte[] values) {
            return new ObjectIdAndroid(values);
        }

    };

    /**
     * Constructor from {@link BluetoothGattCharacteristic}
     *
     * @param bluetoothGattCharacteristic Characteristics UUID: 0x2AC3
     */
    @Deprecated
    public ObjectIdAndroid(@NonNull BluetoothGattCharacteristic bluetoothGattCharacteristic) {
        super(bluetoothGattCharacteristic.getValue());
    }

    /**
     * Constructor from byte array
     *
     * @param values byte array from <a href="https://developer.android.com/reference/android/bluetooth/BluetoothGattCharacteristic#getValue()">BluetoothGattCharacteristic#getValue()</a>
     */
    public ObjectIdAndroid(@NonNull byte[] values) {
        super(values);
    }

    /**
     * Constructor from parameters
     *
     * @param objectId Object ID
     */
    public ObjectIdAndroid(long objectId) {
        super(objectId);
    }

    /**
     * Constructor from {@link Parcel}
     *
     * @param in Parcel
     */
    private ObjectIdAndroid(@NonNull Parcel in) {
        super(Objects.requireNonNull(in.createByteArray()));
    }

    /**
     * {@inheritDoc}
     */
    @Override
    public int describeContents() {
        return 0;
    }

    /**
     * {@inheritDoc}
     */
    @Override
    public void writeToParcel(@NonNull Parcel dest, int flags) {
        dest.writeByteArray(getBytes());
    }

}
